Use enriched `depiction` and `sameAs` data for details view
Remove EntityFacts object and additional index lookup

See https://github.com/hbz/lobid-gnd/issues/69
