--- app/models/EntityFacts.java
@@ -1,187 +1,0 @@
-package models;
-
-import java.io.UnsupportedEncodingException;
-import java.net.URLDecoder;
-import java.nio.charset.StandardCharsets;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.CompletionStage;
-import java.util.concurrent.ExecutionException;
-import java.util.stream.Collectors;
-
-import com.fasterxml.jackson.databind.JsonNode;
-
-import controllers.HomeController;
-import play.Logger;
-import play.libs.Json;
-import play.libs.ws.WSClient;
-import play.libs.ws.WSResponse;
-import play.mvc.Http;
-
-/**
- * Representation of an EntityFacts entity.
- * 
- * See http://hub.culturegraph.org/entityfacts
- * 
- * @author Fabian Steeg (fsteeg)
- *
- */
-public class EntityFacts {
-
-	public List<Map<String, Object>> sameAs;
-	public Map<String, Object> depiction;
-	public String imageAttribution;
-
-	public static EntityFacts entity(WSClient client, String id) {
-		EntityFacts entity = Json.fromJson(json(client, id), EntityFacts.class);
-		if (entity.getImage().url.contains("File:"))
-			entity.imageAttribution = attribution(client,
-					entity.getImage().url.substring(entity.getImage().url.indexOf("File:") + 5).split("\\?")[0]);
-		return entity;
-	}
-
-	public List<Link> getLinks() {
-		return sameAs == null ? Collections.emptyList() : sameAs.stream().map(map -> {
-			String url = map.get("@id").toString();
-			Object icon = null;
-			Object label = null;
-			Object collection = map.get("collection");
-			if (collection != null) {
-				@SuppressWarnings("unchecked")
-				Map<String, Object> collectionMap = (Map<String, Object>) collection;
-				icon = collectionMap.get("icon");
-				label = collectionMap.get("name");
-			}
-			return new Link(url, icon == null ? "" : icon.toString(), label == null ? "" : label.toString());
-		}).collect(Collectors.toList());
-	}
-
-	public Link getImage() {
-		if (depiction != null) {
-			String url = depiction.get("url").toString();
-			String image = depiction.get("@id").toString();
-			Object thumbnailObject = depiction.get("thumbnail");
-			if (thumbnailObject != null) {
-				@SuppressWarnings("unchecked")
-				String thumbnailString = ((Map<String, Object>) thumbnailObject).get("@id").toString();
-				image = thumbnailString;
-			}
-			return new Link(url, image, imageAttribution != null ? imageAttribution : url);
-		}
-		return new Link("", "", ""); // TODO: null better?
-	}
-
-	public static class Link implements Comparable<Link> {
-		public String url;
-		public String image;
-		public String label;
-
-		Link(String url, String icon, String label) {
-			this.url = url;
-			this.image = icon;
-			this.label = label.replace("Gemeinsame Normdatei (GND) im Katalog der Deutschen Nationalbibliothek",
-					"Deutsche Nationalbibliothek (DNB)");
-		}
-
-		@Override
-		public int hashCode() {
-			return url.hashCode();
-		}
-
-		@Override
-		public boolean equals(Object obj) {
-			if (this == obj)
-				return true;
-			if (obj == null)
-				return false;
-			if (getClass() != obj.getClass())
-				return false;
-			Link that = (Link) obj;
-			return that.url.equals(this.url);
-		}
-
-		@Override
-		public String toString() {
-			return "Link [url=" + url + ", icon=" + image + ", label=" + label + "]";
-		}
-
-		@Override
-		public int compareTo(Link that) {
-			return that.url.compareTo(this.url);
-		}
-	}
-
-	private static final String URI_PREFIX = "http://d-nb.info/gnd/";
-
-	private static String attribution(WSClient client, String url) {
-		try {
-			return requestInfo(client, url).thenApply(info -> {
-				String attribution = createAttribution(url, info.asJson());
-				return String.format("Bildquelle: %s", attribution);
-			}).toCompletableFuture().get();
-		} catch (UnsupportedEncodingException | InterruptedException | ExecutionException e) {
-			e.printStackTrace();
-		}
-		return null;
-	}
-
-	private static CompletionStage<WSResponse> requestInfo(WSClient client, String imageName)
-			throws UnsupportedEncodingException {
-		String imageId = "File:" + URLDecoder.decode(imageName, StandardCharsets.UTF_8.name());
-		return client.url("https://commons.wikimedia.org/w/api.php")//
-				.addQueryParameter("action", "query")//
-				.addQueryParameter("format", "json")//
-				.addQueryParameter("prop", "imageinfo")//
-				.addQueryParameter("iiprop", "extmetadata")//
-				.addQueryParameter("titles", imageId).get();
-	}
-
-	private static String createAttribution(String fileName, JsonNode info) {
-		String artist = findText(info, "Artist");
-		String licenseText = findText(info, "LicenseShortName");
-		String licenseUrl = findText(info, "LicenseUrl");
-		String fileSourceUrl = "https://commons.wikimedia.org/wiki/File:" + fileName;
-		return String.format(
-				(artist.isEmpty() ? "%s" : "%s | ") + "<a href='%s'>Wikimedia Commons</a> | <a href='%s'>%s</a>",
-				artist, fileSourceUrl, licenseUrl.isEmpty() ? fileSourceUrl : licenseUrl, licenseText);
-	}
-
-	private static String findText(JsonNode info, String field) {
-		JsonNode node = info.findValue(field);
-		return node != null ? node.get("value").asText().replace("\n", " ").trim() : "";
-	}
-
-	public static JsonNode json(WSClient client, String id) {
-		String gndNumber = id.replace(URI_PREFIX, "");
-		try {
-			JsonNode result = client//
-					.url(String.format(HomeController.config("entityfacts.index"), gndNumber))//
-					.setFollowRedirects(true).get().thenApply(r -> {
-						if (r.getStatus() == Http.Status.NOT_FOUND) {
-							Logger.debug("No result for GND {}", gndNumber);
-						}
-						return r.asJson();
-					}).exceptionally(e -> {
-						Logger.warn("Could not load data from internal entityfacts index ({}), trying live",
-								e.getMessage());
-						return fallback(client, gndNumber);
-					}).toCompletableFuture().get();
-			return result;
-		} catch (InterruptedException | ExecutionException e) {
-			Logger.warn("Could not load additional information form EntityFacts: {}", e.getMessage());
-		}
-		return Json.newObject();
-	}
-
-	private static JsonNode fallback(WSClient client, String gndNumber) {
-		try {
-			String url = HomeController.config("entityfacts.live") + gndNumber;
-			return client.url(url).get().toCompletableFuture().get().asJson();
-		} catch (InterruptedException | ExecutionException e) {
-			Logger.warn("EntityFacts live fallback failed: {}", e.getMessage());
-		}
-		return Json.newObject();
-	}
-
-}
